Rename research API to unified entry point
